backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "mediapipe>=0.10.33",
    "numpy>=2.4.4",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.2.0",
    "python-multipart>=0.0.22",
    "slowapi>=0.1.9",
    "uvicorn[standard]>=0.42.0",
]
